+++ pkg/R/calculators.R
@@ -199,23 +199,35 @@ ratio <- function(i, type = "of.sum", as_density = FALSE, sort.type = c("none", 
 	#' @family Calculators
 	#'
 	#' @examples
-	#' # Argument "sort.type" = "none" (the default):
-	#' y <- ratio(print(x), type = pareto, as_density = FALSE)
-	#' plot(x = x, y = y, col = "blue", main = "cumulative")
-	#' y <- ratio(x, type = pareto, as_density = TRUE) 
-	#' plot(x = x, y = y, col = "red", main = "cumulative w/ density")
-	#' y <- ratio(x, type = of.max, as_density = FALSE) 
-	#' plot(x = x, y = y, col = "green", main = "of.max")
-	#' y <- ratio(x, type = of.max, as_density = TRUE)
-	#' plot(x = x, y = y, col = "orange", main = "of.max w/density")
-	#' # Argument "sort.type" (type = "of.sum", the default"):
-	#' (\(x2) list(
-	#' 	orig = x2
-	#' 	, none = ratio(x2, decimals = 8, sort.type = "none")
-	#' 	, value = ratio(x2, decimals = 8, sort.type = "value")
-	#' 	, label = ratio(x2, decimals = 8, sort.type = "label")
-	#' 	, num_label = ratio(x2, decimals = 8, sort.type = "num_label")
-	#' 	))(sample(100, 5, TRUE) |> (\(m) rlang::set_names(m, m + 50))())
+	#'  (x <- sample(-100:100, 30))
+	#' 
+	#'  (d <- cbind(
+	#'  	x
+	#'  	, of.sum = ratio(x, of.sum)
+	#'  	, of.sum.dens = ratio(x, of.sum, as = TRUE)
+	#'  	, of.max = ratio(x, of.max)
+	#'  	, of.max.dens = ratio(x, of.max, as = TRUE)
+	#'  	, cumulative = ratio(x, cumulative)
+	#'  	, cumulative.dens = ratio(x, cumulative, as = TRUE)
+	#'  	))
+	#' 
+	#'  plot(of.sum ~ x, data = d, main = "of.sum", col = "red")
+	#'  plot(of.max ~ x, data = d, main = "of.max", col = "green")
+	#'  plot(cumulative ~ x, data = d, main = "cumulative", col = "blue")
+	#'  plot(of.sum.dens ~ x, data = d, main = "of.sum w/ density", col = "red")
+	#'  points(of.max.dens ~ x, data = d, main = "of.max w/ density", col = "green")
+	#'  points(cumulative.dens ~ x, data = d, main = "cumulative w/ density", col = "blue")
+	#' 
+	#'  # sort.type:
+	#'  (x2 <- sample(100, 5, TRUE) %>% rlang::set_names(. + 50))
+	#' 
+	#'  list(
+	#'  	`orig := table(...)` = x2
+	#'  	, `ratio(<args>, sort.type = none)` = ratio(x2, sort.type = "none")
+	#'  	, `ratio(<args>, sort.type = value)` = ratio(x2, sort.type = "value")
+	#'  	, `ratio(<args>, sort.type = label)` = ratio(x2, sort.type = "label")
+	#'  	, `ratio(<args>, sort.type = num_label)` = ratio(x2, sort.type = "num_label")
+	#'  	)
 	#' 
 	#' @export
 
@@ -225,11 +237,17 @@ ratio <- function(i, type = "of.sum", as_density = FALSE, sort.type = c("none", 
 	if (is.call(type)){ type <- rlang::call_args(type) }
 	type <- as.character(type)
 
-	# ... legacy code ...
+	# 'type'
 	if (type == "pareto"){ type <- "cumulative" }
+	if (any(is.na(i)) & (type == "cumulative")){
+		cli::cli_abort("'type' is cumulative but NA values detected.")
+	}
 
 	# 'sort.type':
 	sort.type <- match.arg(sort.type)
+
+	# Capture the names of the original vector `i`:
+	names.i <- names(i)
 
 	# `.i` aliases `i` in order to preserve the input:
 	.i <- i
@@ -237,12 +255,7 @@ ratio <- function(i, type = "of.sum", as_density = FALSE, sort.type = c("none", 
 	# Set the origin of the vector to a minimum of zero because ratio levels of measurement require an absolute zero point. 
 	if (any(.i < 0)){ .i <- .i + abs(min(.i)) }
 
-	# Sort `.i` if type is "cumulative":
-	if (type == "cumulative"){ .i %<>% sort() }
-
-	# Capture the names of the original vector `i`:
-	names.i <- names(i)
-
+	# Define actions based on `sort.type`:
 	sort.actions <- rlang::exprs(
 		none = seq_along(i)
 		, value = order(i)
@@ -250,15 +263,15 @@ ratio <- function(i, type = "of.sum", as_density = FALSE, sort.type = c("none", 
 		, num_label = order(as.numeric(names.i))
 		)
 
-	if (!any(names.i == "")){
-		if (sort.type == "num_label"){ 
-			if (rlang::is_empty(names.i)){ 
-				sort.type <- "value"
-			} else if (!all(is.numeric(as.numeric(names.i)))){
-				sort.type <- "label"
+			if (!any(names.i == "")){
+				if (sort.type == "num_label"){ 
+					if (rlang::is_empty(names.i)){ 
+						sort.type <- "none"
+					} else if (!all(is.numeric(as.numeric(names.i)))){
+						sort.type <- "label"
+					}
+				}
 			}
-		}
-	}
 	
 	ord.i <- eval(sort.actions[[sort.type]])
 
@@ -272,15 +285,10 @@ ratio <- function(i, type = "of.sum", as_density = FALSE, sort.type = c("none", 
 	if (as_density){ .out %<>% dens(.type = type) }
 
 	# Update the result with the ordering specified by "sort.type":
-	if (type == "cumulative"){
-		.out <- .out[order(i)]
-	}
 	.out <- .out[ord.i]
 
 	# Restore names if they exist in the input:
-	if (!rlang::is_empty(names.i)){
-		.out <- rlang::set_names(.out, names.i)
-	}
+	if (!rlang::is_empty(names.i)){ .out <- rlang::set_names(.out, names.i[ord.i]) }
 
 	# Return
 	return(.out)